tests: attempting to fix flaky tests during unstable cluster
